+++ src/main/java/enterprise/test/common/batch/KeyCache.java
@@ -1,0 +1,6 @@
+package enterprise.test.common.batch;
+
+import enterprise.test.common.redis.ReviewCache;
+
+public record KeyCache(Long key, ReviewCache cache) {
+}
