+++ src/main/java/enterprise/test/common/redis/ReviewCache.java
@@ -1,4 +1,4 @@
-package enterprise.test.infra.redis;
+package enterprise.test.common.redis;
 
 import enterprise.test.review.domain.Review;
