+++ src/main/java/enterprise/test/common/redis/ReviewCacheRedis.java
@@ -1,12 +1,10 @@
-package enterprise.test.infra.redis;
+package enterprise.test.common.redis;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.dao.DataAccessException;
-import org.springframework.data.redis.core.RedisOperations;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.SessionCallback;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -57,7 +55,6 @@ public class ReviewCacheRedis implements RedisService<ReviewCache, Long> {
 
         int count = Integer.parseInt((String) map.get("count"));
         int sum = Integer.parseInt((String) map.get("sum"));
-        log.info("count: {}, sum: {}", count, sum);
         return new ReviewCache(count, sum);
     }
 
@@ -68,7 +65,6 @@ public class ReviewCacheRedis implements RedisService<ReviewCache, Long> {
     @Override
     public Set<Long> keys() {
         Set<String> keys = redisTemplate.keys(PREFIX + "*");
-        log.info("keys: {}", keys);
 
         if (keys == null || keys.isEmpty()) {
             return Collections.emptySet();
@@ -80,27 +76,11 @@ public class ReviewCacheRedis implements RedisService<ReviewCache, Long> {
     }
 
     @Override
+    @Transactional
     public ReviewCache getAndRemoveValue(Long key) {
-        // 트랜잭션 내에서 실행할 로직을 SessionCallback으로 작성
         ReviewCache cache = getValue(key);
-        return redisTemplate.execute(new SessionCallback<ReviewCache>() {
-            @Override
-            @SuppressWarnings("unchecked")
-            public ReviewCache execute(RedisOperations operations) throws DataAccessException {
-
-                try {
-                    operations.multi(); // 트랜잭션 시작
-                    deleteValue(key); // 값을 삭제하는 명령을 트랜잭션에 추가
-                    operations.exec(); // 트랜잭션 종료
-
-                } catch (Exception e) {
-                    e.printStackTrace();
-                    setValue(key, cache); // 트랜잭션 실패 시 원래 값을 복구
-                    operations.discard();
-                }
-                return cache;
-            }
-        });
+        deleteValue(key);
+        return cache;
     }
 
     @Override
