+++ src/main/java/enterprise/test/common/time/CustomClock.java
@@ -1,4 +1,4 @@
-package enterprise.test.common;
+package enterprise.test.common.time;
 
 public interface CustomClock<T> {
