--- src/main/java/enterprise/test/infra/batch/KeyCache.java
@@ -1,6 +1,0 @@
-package enterprise.test.infra.batch;
-
-import enterprise.test.infra.redis.ReviewCache;
-
-public record KeyCache(Long key, ReviewCache cache) {
-}
